requirements.txt
certifi==2021.10.8
numpy==1.22.2
Pillow==9.0.1
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
typing_extensions==4.1.1
